Cleaned up graceful shutdown messages, updated function comments
Since graceful shutdown can be performed in more than one spot, tried to make it flow a bit better by factoring out the messages that can be used. Additionally, noticed that the function comments were not up to date, so I went ahead and checked them all in the file I was working in.