Added first unit test
Had to change what work was done in the Simple Slack Bot initializer by introducing a new method called `connect()` that needs to be called after initialization. This is so we can write our unit tests and perform dependency injection. Our first unit test was testing that registration actually resulted in appending a callback to the approrpriate list, using the hello Slack API event.