Added Old Unit Tests To Master (#40)
* Added first unit test

Had to change what work was done in the Simple Slack Bot initializer by introducing a new method called `connect()` that needs to be called after initialization. This is so we can write our unit tests and perform dependency injection. Our first unit test was testing that registration actually resulted in appending a callback to the approrpriate list, using the hello Slack API event.

* Added coverage configuration so we can generate test code coverage

* Updated readme with how to run unit tests and generate coverage

* Woops actually updated README.md

* Tested register function with all Slack Bot RTM events.

This was achieved by adding the ability to write a unit test with a param, so we could exhaustively try all events.

* wrote test for route_request_to_callbacks

* Attempt at setting up Circle CI to automatically run our unit tests

* Woops forgot our circleci config and directory

* removed unused uploading of artifacts

* Got test for getSlacker working. First non-parameterized test along side parameterized ones

* Added assert messages

* Added test for get_slack_socket

* Added None cases to getters

* Wrote test for helper_get_public_channel_ids and found potential bug if Slacker wasn't set, fixed it

* WIP writing mock for Slacker so we can further test helper_get_private_channel_ids. Probably should mock out other dependencies like SlackRequest and SlackEvent

* commented out broken test

* merging master to this old thread

* fix: unit tests so they actually pass now, specifically helper function private channels

Had to get mockin correctly. These tests are not pretty but I want the
green checkmark