Remove Slacker and replace with Python SlackClient
Since Slacker is now deprecated, we should be moving towards the correct
way to interact with the Slack API. This is a rather lengthy commit,
larger than I had hoped, but it removes Slacker for Python SlackClient.

I have not successfully ran the ExampleComponent, specifically it is not
writing to the channel it is in. I'll be investigating that before
submitting this branch as a PR.