Replace slacker with python slackclient (#44)
Remove Slacker and replace with Python SlackClient

Since Slacker is now deprecated, we should be moving towards the correct
way to interact with the Slack API. This is a rather lengthy commit,
larger than I had hoped, but it removes Slacker for Python SlackClient.

I have not successfully ran the ExampleComponent, specifically it is not
writing to the channel it is in. I'll be investigating that before
submitting this branch as a PR.

Fixed example_component to be working with new python_slackclient

Completly removed Slacker from dependencies. The "hello" event one
receives from Slack does not dictate a channel, so I removed that from
the example_component. Additionally, I was getting rate limited when
testing the example_component due to responding to a user typing, so I
also removed that.

Replaced unittest with pytest and overhauled libraries used and Makefile.

Ended up adding a ton of static analyzers and other libraries to help improve the quality of the code. Improved the circleci config as well and added a badge to the README.md