Update documentation (#49)
Moved the Ping Pong Bot example to a new examples directory. Can later add more examples of using the framework. Additionally dockerized this example for ease of productionifying (is that a word?)

Additionally moved our dependency versions up as an old version was failing to interact with the Slack API under the hood.

Got a little distracted in mostly developing a `$ make deploy` command to make it easier to deploy a new framework update. This command is failing to create a new Github Release when hitting the API, specifically returning 

> Failed to create Github Release for version v2.3.2. Output was b'{"message":"Not Found","documentation_url":"https://docs.github.com/rest/reference/repos#create-a-release"}'

I'll fix this in another PR